requirements.txt
scrapy
jikanpy-v4
fuzzywuzzy[speedup]